remove redundant QueryTest
